Switched from double to float.

+++ matrix.c
@@ -3,7 +3,7 @@
 vector create_vector(size_t size) {
 	vector ret;
 	ret.size = size;
-	ret.data = checked_malloc2(size, sizeof(double));
+	ret.data = checked_malloc2(size, sizeof(float));
 	return ret;
 }
 
@@ -14,7 +14,7 @@ void free_vector(vector vec) {
 matrix create_matrix(size_t size) {
 	matrix ret;
 	ret.size = size;
-	ret.data = checked_malloc3(size, size, sizeof(double));
+	ret.data = checked_malloc3(size, size, sizeof(float));
 	return ret;
 }
 
@@ -29,7 +29,7 @@ void matrix_vector_mul(matrix m, vector v, vector result) {
 	
 	size_t pos = 0;
 	for(size_t i = 0; i < n; ++i) {
-		double val = 0.0;
+		float val = 0.0;
 		for(size_t j = 0; j < n; ++j) {
 			val += m.data[pos] * v.data[j];
 			++pos;
