+++ vec3.h
@@ -6,7 +6,7 @@
 /// Structure representing one vector in 3D.
 typedef struct {
 	/// The coordinates of the vector.
-	double x, y, z;
+	float x, y, z;
 } vec3;
 
 /// Returns the sum of vectors \p a and \p b.
@@ -37,7 +37,7 @@ static inline vec3 vec3_neg(vec3 v) {
 }
 
 /// Returns the vector \p v multiplied by scalar \p c.
-static inline vec3 vec3_mul(vec3 v, double c) {
+static inline vec3 vec3_mul(vec3 v, float c) {
 	v.x *= c;
 	v.y *= c;
 	v.z *= c;
@@ -54,15 +54,15 @@ static inline vec3 vec3_cross(vec3 a, vec3 b) {
 }
 
 /// Computes the dot product of vectors \p a and \p b.
-static inline double vec3_dot(vec3 a, vec3 b) {
+static inline float vec3_dot(vec3 a, vec3 b) {
 	return a.x * b.x + a.y * b.y + a.z * b.z;
 }
 
 /// Computes the square of the length of vector \p v.
-static inline double vec3_len2(vec3 v) {
+static inline float vec3_len2(vec3 v) {
 	return v.x * v.x + v.y * v.y + v.z * v.z;
 }
 /// Computes the length of vector \p v.
-static inline double vec3_len(vec3 v) {
+static inline float vec3_len(vec3 v) {
 	return sqrt(vec3_len2(v));
 }
